Update statements.go
Int32 is marshalled in json, used the same datatype in clickhouse to avoid potential errors

+++ client/pkg/clickhouse/statements.go
@@ -153,7 +153,7 @@ const azureContainerPushEventTable DBStatement = `
 		ImageName String,
 		Event String,
 		Timestamp String,
-		Size Int64,
+		Size Int32,
 		SHAID String
 	) engine=File(TabSeparated)
 	`
